requirements.txt
beautifulsoup4==4.14.3
fastapi==0.135.1
httpx==0.28.1
psycopg2-binary==2.9.11
pytest==9.0.2
requests==2.32.5
uvicorn==0.41.0

